apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.13
pydantic-settings==2.6.1
pytest==8.3.3
httpx==0.27.2
ruff==0.7.4
